App/requirements.txt
pyside6
qasync
simpleobsws
pyserial